Allow decimal tips and withdrawals

+++ command.rb
@@ -69,7 +69,7 @@ class Command
 
   def set_amount
     amount = @params.shift
-    @amount = amount.to_i
+    @amount = amount.to_f
     randomize_amount if (amount == "random")
 
     raise @coin_config_module::TOO_POOR_TEXT unless available_balance >= @amount + 0.1
